+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Support for PHP 8.0 is dropped. Minimal PHP version required is 8.2.
 
 Enhancements:
 * [ORM] Added support for count query with `DISTINCT` in PostgreSQL dialect.
+* [Utilities] `Strings::EMPTY_STRING` is deprecated in favour of `Strings::EMPTY`. 
 
 Release 2.0.0
 --------